refactoring: combining all Header-related classes to one class (HeaderTest)
