+++ src/test/java/school/redrover/HeaderTest.java
@@ -1,37 +1,19 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
+import org.openqa.selenium.JavascriptExecutor;
+import org.openqa.selenium.Keys;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.interactions.Actions;
 import org.testng.Assert;
 import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.runner.BaseTest;
-
-import java.util.ArrayList;
 import java.util.List;
 
+import static org.testng.Assert.assertTrue;
+
 public class HeaderTest extends BaseTest {
-
-    @Test
-    public void testReturnWithLogo() {
-        getDriver().findElement(By.xpath("//a[@href='/manage']")).click();
-
-        getDriver().findElement(By.xpath("//img[@id='jenkins-name-icon']")).click();
-
-        Assert.assertTrue(getDriver().getTitle().contains("Dashboard"));
-    }
-
-    @Ignore
-    @Test
-    public void testClickLogoToMainPage() {
-
-        getDriver().findElement(By.xpath("//a[@href='/me/my-views']")).click();
-
-        getDriver().findElement(By.id("jenkins-head-icon")).click();
-
-        Assert.assertTrue(
-                getDriver().findElement(By.cssSelector(".empty-state-block > h1")).getText().contains("Welcome to Jenkins!"));
-    }
-
     @Test
     public void testReturningBackToMainPageFromMainMenuPages() {
 
@@ -49,11 +31,94 @@ public class HeaderTest extends BaseTest {
             Assert.assertEquals(getDriver().getTitle(), "Dashboard [Jenkins]");
         }
     }
+    private void goToJenkinsHomePage() {
+        getDriver().findElement(By.id("jenkins-name-icon")).click();
+    }
+
+    private void createFreeStyleProject(String projectName) {
+        goToJenkinsHomePage();
+        getDriver().findElement(By.linkText("New Item")).click();
+        getDriver().findElement(By.className("hudson_model_FreeStyleProject")).click();
+        getDriver().findElement(By.id("name")).sendKeys(projectName);
+        getDriver().findElement(By.id("ok-button")).click();
+    }
 
     @Test
-    public void testVerifyRedirectToHomepageByClickLogoImg() {
-        getDriver().findElement(By.id("jenkins-head-icon")).click();
+    public void testExactMatchSearchFunctionality() {
+        final String itemName = "Test project";
+        createFreeStyleProject(itemName);
 
-        Assert.assertEquals(getDriver().getTitle().trim(), "Dashboard [Jenkins]");
+        getDriver().findElement(By.name("q")).click();
+        getDriver().findElement(By.name("q")).sendKeys(itemName);
+        new Actions(getDriver()).sendKeys(Keys.ENTER).perform();
+
+        String title = getDriver().getTitle();
+        boolean isStatusPageSelected = getDriver()
+                .findElement(By.linkText("Status"))
+                .getAttribute("class")
+                .contains("active");
+
+        assertTrue(title.contains(itemName));
+        assertTrue(isStatusPageSelected);
     }
+
+    @Test
+    public void testPartialMatchSearchFunctionality() {
+        final String itemName1 = "Test project1";
+        final String itemName2 = "Test project2";
+        final String itemName3 = "Test project3";
+        final String searchingRequest = itemName1.substring(0, 5);
+        createFreeStyleProject(itemName1);
+        createFreeStyleProject(itemName2);
+        createFreeStyleProject(itemName3);
+
+        getDriver().findElement(By.name("q")).click();
+        getDriver().findElement(By.name("q")).sendKeys(searchingRequest);
+        new Actions(getDriver()).sendKeys(Keys.ENTER).perform();
+
+        boolean isResultMatchQuery = getDriver()
+                .findElements(By.xpath("//div[@id='main-panel']//li"))
+                .stream()
+                .map(WebElement::getText)
+                .allMatch(x -> x.contains(searchingRequest));
+        assertTrue(isResultMatchQuery);
+    }
+
+    @Test
+    public void testRedirectionToStatusPageFromResultList() {
+        final String itemName = "Test project";
+        final String searchRequest = itemName.substring(0, 5);
+        createFreeStyleProject(itemName);
+
+        getDriver().findElement(By.name("q")).click();
+        getDriver().findElement(By.name("q")).sendKeys(searchRequest);
+        new Actions(getDriver()).sendKeys(Keys.ENTER).perform();
+        getDriver().findElement(By.linkText(itemName)).click();
+
+        String title = getDriver().getTitle();
+        boolean isStatusPageSelected = getDriver()
+                .findElement(By.linkText("Status"))
+                .getAttribute("class")
+                .contains("active");
+
+        assertTrue(title.contains(itemName));
+        assertTrue(isStatusPageSelected);
+    }
+
+    @Ignore
+    @Test
+    public void testHotKeysSearchAreaSelection() {
+        new Actions(getDriver())
+                .keyDown(Keys.CONTROL)
+                .sendKeys("k")
+                .keyUp(Keys.CONTROL)
+                .perform();
+
+        boolean isFocused = (Boolean) ((JavascriptExecutor) getDriver()).executeScript(
+                "return document.activeElement === arguments[0]", getDriver().findElement(By.name("q")));
+        assertTrue(isFocused);
+    }
+
+
+
 }
